flaskProject/requirements.txt
boto3==1.26.94
click==8.1.3
colorama==0.4.6
filetype==1.2.0
Flask==2.2.3
Flask-Cors==3.0.10
Flask-SQLAlchemy==3.0.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
pip==23.0.1
psycopg2==2.9.5
python-dotenv==1.0.0
setuptools==67.6.0
Werkzeug==2.2.3
wheel==0.40.0
